@@ -9,10 +9,10 @@
   },
   "homepage": "https://appbaseio-apps.github.io/gitxplore-app",
   "dependencies": {
-    "@appbaseio/reactivesearch": "^2.3.0",
-    "react": "^16.2.0",
+    "@appbaseio/reactivesearch": "^3.0.0-rc.12",
+    "react": "^16.8.0",
     "react-dom": "^16.2.0",
-    "react-scripts": "1.0.17"
+    "react-scripts": "3.0.1"
   },
   "scripts": {
     "start": "react-scripts start",
@@ -24,6 +24,18 @@
   },
   "license": "MIT",
   "devDependencies": {
-    "gh-pages": "^1.1.0"
+    "gh-pages": "^2.0.1"
+  },
+  "browserslist": {
+    "production": [
+      ">0.2%",
+      "not dead",
+      "not op_mini all"
+    ],
+    "development": [
+      "last 1 chrome version",
+      "last 1 firefox version",
+      "last 1 safari version"
+    ]
   }
 }
