@@ -5,6 +5,7 @@ import {
 	SingleDropdownRange,
 	RangeSlider,
 } from '@appbaseio/reactivesearch';
+import { repoActiveDropdownData, repoCreatedDropdownData } from './constants';
 
 const SearchFilters = ({ currentTopics, setTopics, visible }) => (
 	<div className={`flex column filters-container ${!visible ? 'hidden' : ''}`}>
@@ -37,11 +38,7 @@ const SearchFilters = ({ currentTopics, setTopics, visible }) => (
 				placeholder="Repo last active"
 				title="Last Active"
 				filterLabel="Last Active"
-				data={[
-					{ start: 'now-1M', end: 'now', label: 'Last 30 days' },
-					{ start: 'now-6M', end: 'now', label: 'Last 6 months' },
-					{ start: 'now-1y', end: 'now', label: 'Last year' },
-				]}
+				data={repoActiveDropdownData}
 			/>
 		</div>
 		<div className="child m10">
@@ -51,63 +48,7 @@ const SearchFilters = ({ currentTopics, setTopics, visible }) => (
 				placeholder="Repo created"
 				title="Created"
 				filterLabel="Created"
-				data={[
-					{
-						start: '2017-01-01T00:00:00Z',
-						end: '2017-12-31T23:59:59Z',
-						label: '2017',
-					},
-					{
-						start: '2016-01-01T00:00:00Z',
-						end: '2016-12-31T23:59:59Z',
-						label: '2016',
-					},
-					{
-						start: '2015-01-01T00:00:00Z',
-						end: '2015-12-31T23:59:59Z',
-						label: '2015',
-					},
-					{
-						start: '2014-01-01T00:00:00Z',
-						end: '2014-12-31T23:59:59Z',
-						label: '2014',
-					},
-					{
-						start: '2013-01-01T00:00:00Z',
-						end: '2013-12-31T23:59:59Z',
-						label: '2013',
-					},
-					{
-						start: '2012-01-01T00:00:00Z',
-						end: '2012-12-31T23:59:59Z',
-						label: '2012',
-					},
-					{
-						start: '2011-01-01T00:00:00Z',
-						end: '2011-12-31T23:59:59Z',
-						label: '2011',
-					},
-					{
-						start: '2010-01-01T00:00:00Z',
-						end: '2010-12-31T23:59:59Z',
-						label: '2010',
-					},
-					{
-						start: '2009-01-01T00:00:00Z',
-						end: '2009-12-31T23:59:59Z',
-						label: '2009',
-					},
-					{
-						start: '2008-01-01T00:00:00Z',
-						end: '2008-12-31T23:59:59Z',
-						label: '2008',
-					},
-					{
-						start: '2007-01-01T00:00:00Z',
-						end: '2007-12-31T23:59:59Z',
-						label: '2007',
-					},
-				]}
+				data={repoCreatedDropdownData}
 			/>
 		</div>
 		<div className="child m10">
